Alterando router
