+++ pokedex2/src/components/CardPokemon/CardPokemon.js
