+++ pokedex2/src/components/CardPokemon/StyledCardPokemon.js
