+++ pokedex2/src/components/CardPokemon/StyledCardsHome.js
