+++ pokedex2/src/images/errorgif.gif
